requirements.txt
# Core dependencies
torch>=2.1.0
transformers>=4.36.0
protobuf>=4.25.1

# Web application
flask>=3.0.0
werkzeug>=3.0.1

# Document processing
python-docx>=1.0.1
PyPDF2>=3.0.0

# Text processing
nltk>=3.8.1
textstat>=0.7.3
spacy>=3.7.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl

# Metrics and evaluation
rouge-score>=0.1.2
bert-score>=0.3.13
numpy>=1.24.3
scikit-learn>=1.3.2
imbalanced-learn>=0.11.0

# Utilities
tqdm>=4.66.1
pandas>=2.1.4
pyyaml>=6.0.1
python-dotenv>=1.0.0
requests>=2.31.0

# Performance optimization
accelerate>=0.25.0
bitsandbytes>=0.41.1
safetensors>=0.4.1 